Make sure there are no concrete nodes that are used as base classes
